+++ src/main/java/com/jorgedirkx/repositories/BugRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface BugRepository extends CrudRepository<Bug,Long> {
+public interface BugRepository extends CrudRepository<Bug,Integer> {
 
     /*List<Bug> findByName(String name);*/
 }
